Much refactoring
