+++ .dockerignore
@@ -1,0 +1,7 @@
+**/node_modules
+**/npm-debug.log
+**/.DS_Store
+**/default.env
+**/.env
+**/request.json
+**/response.json
