+++ node-docker/.dockerignore
@@ -4,4 +4,3 @@ node_modules
 npm-debug.log
 .DS_Store
 default.env
-.env
