+++ node-docker/default.env
@@ -1,0 +1,10 @@
+BUCKET_ENDPOINT=s3.REGION_NAME.backblazeb2.com
+BUCKET_NAME=
+ACCESS_KEY=
+SECRET_KEY=
+FRAMEIO_TOKEN=fio-XXXX
+FRAMEIO_SECRET=
+UPLOAD_PATH=fio_exports/
+DOWNLOAD_PATH=b2_imports
+QUEUE_SIZE=4
+PART_SIZE=5242880
