+++ node-docker/package.json
@@ -2,7 +2,8 @@
   "name": "b2-frameio",
   "version": "0.0.1",
   "description": "",
-  "main": "index.js",
+  "main": "server.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -13,11 +14,9 @@
     "node": ">=18.2.0"
   },
   "dependencies": {
-    "aws-sdk": "^2.1163.0",
     "backblaze-frameio-common": "file:../backblaze-frameio-common",
     "compression": "^1.7.4",
     "express": "^4.18.1",
-    "http-errors": "^2.0.0",
     "node-fetch": "^2.6.0"
   }
 }
