+++ node-risingcloud/task/.npmrc
@@ -1,0 +1,2 @@
+# risingcloud build doesn't like symlinks
+install-links=true
