+++ node-risingcloud/webservice/.npmrc
@@ -1,0 +1,2 @@
+# risingcloud build doesn't like symlinks
+install-links=true
