+++ node-risingcloud/webservice/package.json
@@ -13,8 +13,9 @@
     "node": ">=18.2.0"
   },
   "dependencies": {
-    "aws-sdk": "^2.1163.0",
-    "node-fetch": "^2.6.0",
-    "backblaze-frameio-common": "file:../../backblaze-frameio-common"
+    "backblaze-frameio-common": "file:../../backblaze-frameio-common",
+    "compression": "^1.7.4",
+    "express": "^4.18.1",
+    "node-fetch": "^2.6.0"
   }
 }
