--- python-walk/Procfile
@@ -1,1 +1,0 @@
-web: uvicorn main:app --host 0.0.0.0 --port $PORT
