--- python-walk/default.env
@@ -1,5 +1,0 @@
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
-S3_BUCKET=
-S3_ENDPOINT=https://s3.<region>.backblazeb2.com
-FIO_TOKEN=
