--- python-walk/requirements.txt
@@ -1,38 +1,0 @@
-analytics-python==1.4.0; python_full_version >= "3.6.5" and python_version < "4"
-ansicon==1.89.0; python_full_version >= "3.6.5" and python_version < "4" and python_version >= "2.7" and platform_system == "Windows"
-asgiref==3.5.0; python_version >= "3.7"
-backoff==1.10.0; python_full_version >= "3.6.5" and python_version < "4"
-blessed==1.19.1; python_full_version >= "3.6.5" and python_version < "4" and python_version >= "2.7"
-boto3==1.21.31; python_version >= "3.6"
-botocore==1.24.31; python_version >= "3.6"
-certifi==2021.10.8; python_full_version >= "3.6.5" and python_version < "4"
-charset-normalizer==2.0.12; python_full_version >= "3.6.5" and python_version < "4" and python_version >= "3"
-click==8.1.2; python_version >= "3.7"
-colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
-enlighten==1.10.2; python_full_version >= "3.6.5" and python_version < "4"
-fastapi==0.68.2; python_full_version >= "3.6.1"
-frameioclient @ git+https://github.com/Frameio/python-frameio-client.git@develop ; python_full_version >= "3.6.5" and python_version < "4"
-gunicorn==20.1.0; python_version >= "3.5"
-h11==0.13.0; python_version >= "3.6"
-idna==3.3; python_full_version >= "3.6.5" and python_version < "4" and python_version >= "3.5"
-jinxed==1.1.0; python_full_version >= "3.6.5" and python_version < "4" and python_version >= "2.7" and platform_system == "Windows"
-jmespath==1.0.0; python_version >= "3.7"
-monotonic==1.6; python_full_version >= "3.6.5" and python_version < "4"
-prefixed==0.3.2; python_full_version >= "3.6.5" and python_version < "4"
-pydantic==1.9.0; python_full_version >= "3.6.1"
-pyngrok==5.1.0; python_version >= "3.5"
-python-dateutil==2.8.2; python_full_version >= "3.6.5" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6")
-python-dotenv==0.19.2; python_version >= "3.5"
-pyyaml==6.0; python_version >= "3.6"
-requests==2.27.1; python_full_version >= "3.6.5" and python_version < "4"
-s3transfer==0.5.2; python_version >= "3.6"
-simplejson==3.17.6; (python_version >= "2.5" and python_full_version < "3.0.0") or (python_full_version >= "3.3.0")
-six==1.16.0; python_full_version >= "3.6.5" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6") and python_version >= "2.7"
-speedtest-cli==2.1.3; python_full_version >= "3.6.5" and python_version < "4"
-starlette==0.14.2; python_version >= "3.6" and python_full_version >= "3.6.1"
-token-bucket==0.3.0; python_full_version >= "3.6.5" and python_version < "4" and python_version >= "3.5"
-typing-extensions==4.1.1; python_version >= "3.6" and python_full_version >= "3.6.1"
-urllib3==1.26.9; python_full_version >= "3.6.5" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6")
-uvicorn==0.15.0
-wcwidth==0.2.5; python_full_version >= "3.6.5" and python_version < "4" and python_version >= "2.7"
-xxhash==3.0.0; python_full_version >= "3.6.5" and python_version < "4" and python_version >= "3.6"
